Switch otel to use environment variables to pickup trace context.
The sidecar is intended to be run in short-lived builder pods which will inherit a trace context from the build orchestrator. This should be set on the builder pod using environment variables that are consistent with the Jenkins Opentelemetry Plugin, here:

https://github.com/jenkinsci/opentelemetry-plugin/blob/master/docs/job-traces.md#environment-variables-for-trace-context-propagation-and-integrations

They will be parsed using the same logic as the W3C trace context standard:

https://www.w3.org/TR/trace-context/#relationship-between-the-headers

To do this, we simply set the quarkus.opentelemetry.propagators configuration to include the 'envar' propagator, which is provided in the new opentelemetry-ext-cli-java dependency.